Added filter to parent services page
